Simplify

+++ src/Qowaiv/IO/StreamSize.cs
@@ -474,7 +474,7 @@ public readonly partial struct StreamSize : IXmlSerializable, IFormattable, IEqu
             var decimalFormat = GetWithoutStreamSizeMarker(format, streamSizeMarker);
             var mp = GetMultiplier(streamSizeMarker);
 
-            decimal size = (decimal)m_Value / (decimal)mp;
+            decimal size = (decimal)m_Value / mp;
 
             return size.ToString(decimalFormat, formatProvider) + streamSizeMarker;
         }
